Created UnitTests for weather.py Methods Ensures correct functionality of tests
